requirements.txt

Django==1.9.5
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.3.3

mock==1.3.0
mongoengine==0.13.0
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2017.2
readability-lxml==0.6.2
requests==2.9.1
Scrapy==1.4.0
scrapyd==1.2.0
six==1.9.0
stringutil==1.1.2
Twisted==15.4.0
